+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `--output`
 - Fix quoted source paths breaking tailwind.config.js syntax
+- Workaround Tailwind not crashing with non-zero exitcode
 
 ## 0.1.0.0 -- 2022-01-08